+++ R/capture-requests.R
@@ -28,6 +28,12 @@
 #' that is written when capturing requests containing the absolute path of the
 #' file. Useful for debugging if you're capturing but don't see the fixture
 #' files being written in the expected location. Default is `FALSE`.
+#' @param redact function to run to purge sensitive strings from the recorded
+#' response objects. It should take a `response`-class object as its argument
+#' and should return a cleansed `response` object. This allows you to remove
+#' certain request and response contents, such as authentication tokens, from
+#' the mocks that get written out. See [redact_auth()], the default, for more
+#' details.
 #' @return `capture_requests` returns the result of `expr`. `start_capturing`
 #' invisibly returns the `path` it is given. `stop_capturing` returns nothing;
 #' it is called for its side effects.
@@ -50,23 +56,28 @@
 #' stop_capturing()
 #' }
 #' @export
-capture_requests <- function (expr, path=.mockPaths()[1], simplify=TRUE, verbose=FALSE) {
-    start_capturing(path, simplify=simplify, verbose=verbose)
+capture_requests <- function (expr, path=.mockPaths()[1], simplify=TRUE,
+                              verbose=FALSE, redact=redact_auth) {
+    start_capturing(path, simplify, verbose, redact)
     on.exit(stop_capturing())
     eval.parent(expr)
 }
 
 #' @rdname capture_requests
 #' @export
-start_capturing <- function (path=.mockPaths()[1], simplify=TRUE, verbose=FALSE) {
+start_capturing <- function (path=.mockPaths()[1], simplify=TRUE, verbose=FALSE,
+                             redact=redact_auth) {
     ## Use "substitute" so that "path" gets inserted. Code remains quoted.
     req_tracer <- substitute({
-        f <- file.path(path, buildMockURL(req))
-        dir.create(dirname(f), showWarnings=FALSE, recursive=TRUE)
-        ## Get the value returned from the function
-        .resp <- returnValue()
+        ## Get the value returned from the function, and sanitize it
+        .resp <- redact(returnValue())
         ## Omit curl handle C pointer
         .resp$handle <- NULL
+
+        ## Construct the mock file path
+        f <- file.path(path, buildMockURL(.resp$request))
+        dir.create(dirname(f), showWarnings=FALSE, recursive=TRUE)
+
         ## Get the Content-Type
         ct <- unlist(headers[tolower(names(headers)) == "content-type"])
         is_json <- any(grepl("application/json", ct))
@@ -94,7 +105,7 @@ start_capturing <- function (path=.mockPaths()[1], simplify=TRUE, verbose=FALSE)
             dput(.resp, file=f)
         }
         if (verbose) message("Writing ", normalizePath(f))
-    }, list(path=path, simplify=simplify, verbose=verbose))
+    }, list(path=path, simplify=simplify, verbose=verbose, redact=redact))
     dl_tracer <- substitute({
         if (status == 0) {
             ## Only do this if the download was successful